Fix VS warnings

+++ include/polar/math/constants.h
@@ -1,7 +1,7 @@
 #pragma once
 
 namespace polar::math {
-	constexpr double PI          = 3.14159265358979323846264338327950288419716939937510;
-	constexpr double TWO_PI      = PI * 2.0;
-	template<size_t DENOM> constexpr double PI_OVER = PI / DENOM;
+	constexpr Decimal PI                             = Decimal(3.14159265358979323846264338327950288419716939937510);
+	constexpr Decimal TWO_PI                         = Decimal(PI * 2.0);
+	template<size_t DENOM> constexpr Decimal PI_OVER = Decimal(PI / DENOM);
 } // namespace polar::math
